requirements.txt
pandas==2.2.3
clickhouse-connect==0.8.15
pyodbc==5.2.0
mysql-connector-python==9.2.0
